infra/setup/requirements.txt
typer>=0.9.0
rich>=13.7.0
pyyaml>=6.0.1
